requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-jose==3.3.0
redis==5.0.8
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
python-multipart==0.0.9
httpx==0.27.2
